+++ execution_engine/src/core/engine_state/mod.rs
@@ -36,7 +36,7 @@ use casper_types::{
     contracts::NamedKeys,
     system::{
         auction::{
-            self, EraValidators, ARG_ERA_END_TIMESTAMP_MILLIS, ARG_EVICTED_VALIDATORS,
+            EraValidators, ARG_ERA_END_TIMESTAMP_MILLIS, ARG_EVICTED_VALIDATORS,
             ARG_REWARD_FACTORS, ARG_VALIDATOR_PUBLIC_KEYS, AUCTION_DELAY_KEY,
             LOCKED_FUNDS_PERIOD_KEY, SEIGNIORAGE_RECIPIENTS_SNAPSHOT_KEY, UNBONDING_DELAY_KEY,
             VALIDATOR_SLOTS_KEY,
@@ -85,6 +85,7 @@ use crate::{
         },
         trie::Trie,
     },
+    system::auction,
 };
 
 /// The maximum amount of motes that payment code execution can cost.
@@ -1757,7 +1758,7 @@ where
             }
         };
 
-        let era_validators_result = auction::era_validators_from_snapshot(snapshot);
+        let era_validators_result = auction::detail::era_validators_from_snapshot(snapshot);
         Ok(era_validators_result)
     }
