+++ execution_engine/src/core/runtime/mod.rs
@@ -30,11 +30,9 @@ use casper_types::{
     },
     system::{
         self,
-        auction::{self, Auction, EraInfo},
-        handle_payment::{self, HandlePayment},
-        mint::{self, Mint},
-        standard_payment::{self, StandardPayment},
-        CallStackElement, SystemContractType, AUCTION, HANDLE_PAYMENT, MINT, STANDARD_PAYMENT,
+        auction::{self, EraInfo},
+        handle_payment, mint, standard_payment, CallStackElement, SystemContractType, AUCTION,
+        HANDLE_PAYMENT, MINT, STANDARD_PAYMENT,
     },
     AccessRights, ApiError, CLTyped, CLValue, ContextAccessRights, ContractHash,
     ContractPackageHash, ContractVersionKey, ContractWasm, DeployHash, EntryPointType, EraId, Gas,
@@ -54,6 +52,10 @@ use crate::{
         wasm_prep::{self, PreprocessingError},
     },
     storage::global_state::StateReader,
+    system::{
+        auction::Auction, handle_payment::HandlePayment, mint::Mint,
+        standard_payment::StandardPayment,
+    },
 };
 pub use stack::{RuntimeStack, RuntimeStackFrame, RuntimeStackOverflow};
