+++ execution_engine/src/core/runtime/standard_payment_internal.rs
@@ -1,14 +1,15 @@
 use casper_types::{
-    system::{
-        handle_payment, mint,
-        standard_payment::{AccountProvider, HandlePaymentProvider, MintProvider, StandardPayment},
-    },
+    system::{handle_payment, mint},
     ApiError, Key, RuntimeArgs, StoredValue, URef, U512,
 };
 
 use crate::{
     core::{execution, runtime::Runtime},
     storage::global_state::StateReader,
+    system::standard_payment::{
+        account_provider::AccountProvider, handle_payment_provider::HandlePaymentProvider,
+        mint_provider::MintProvider, StandardPayment,
+    },
 };
 
 pub(crate) const METHOD_GET_PAYMENT_PURSE: &str = "get_payment_purse";
