+++ execution_engine/src/lib.rs
@@ -18,3 +18,4 @@ pub mod core;
 pub mod shared;
 /// Storage for the execution engine.
 pub mod storage;
+mod system;
