+++ execution_engine/src/system/handle_payment/mint_provider.rs
@@ -1,4 +1,6 @@
-use crate::{account::AccountHash, system::handle_payment::Error, TransferredTo, URef, U512};
+use casper_types::{
+    account::AccountHash, system::handle_payment::Error, TransferredTo, URef, U512,
+};
 
 /// Provides an access to mint.
 pub trait MintProvider {
