+++ execution_engine/src/system/mint/system_provider.rs
@@ -1,4 +1,4 @@
-use crate::{account::AccountHash, system::mint::Error, URef, U512};
+use casper_types::{account::AccountHash, system::mint::Error, URef, U512};
 
 /// Provides functionality of a system module.
 pub trait SystemProvider {
