+++ types/src/system.rs
@@ -1,0 +1,14 @@
+//! System modules, formerly known as "system contracts"
+pub mod auction;
+mod call_stack_element;
+mod error;
+pub mod handle_payment;
+pub mod mint;
+pub mod standard_payment;
+mod system_contract_type;
+
+pub use call_stack_element::CallStackElement;
+pub use error::Error;
+pub use system_contract_type::{
+    SystemContractType, AUCTION, HANDLE_PAYMENT, MINT, STANDARD_PAYMENT,
+};
