+++ types/src/system/auction/bid.rs
