+++ types/src/system/auction/unbonding_purse.rs
@@ -65,9 +65,9 @@ impl UnbondingPurse {
 
     /// Returns public key of unbonder.
     ///
-    /// For withdrawal requests that originated from validator's public key through
-    /// [`crate::system::auction::Auction::withdraw_bid`] entrypoint this is equal to
-    /// [`UnbondingPurse::validator_public_key`] and [`UnbondingPurse::is_validator`] is `true`.
+    /// For withdrawal requests that originated from validator's public key through `withdraw_bid`
+    /// entrypoint this is equal to [`UnbondingPurse::validator_public_key`] and
+    /// [`UnbondingPurse::is_validator`] is `true`.
     pub fn unbonder_public_key(&self) -> &PublicKey {
         &self.unbonder_public_key
     }
@@ -136,7 +136,7 @@ mod tests {
         AccessRights, PublicKey, SecretKey, URef, U512,
     };
 
-    const BONDING_PURSE: URef = URef::new([41; 32], AccessRights::READ_ADD_WRITE);
+    const BONDING_PURSE: URef = URef::new([14; 32], AccessRights::READ_ADD_WRITE);
     const ERA_OF_WITHDRAWAL: EraId = EraId::MAX;
 
     fn validator_public_key() -> PublicKey {
