+++ types/src/system/mint.rs
@@ -1,0 +1,8 @@
+//! Contains implementation of a Mint contract functionality.
+mod constants;
+mod entry_points;
+mod error;
+
+pub use constants::*;
+pub use entry_points::mint_entry_points;
+pub use error::Error;
