+++ types/src/system/standard_payment.rs
@@ -1,0 +1,6 @@
+//! Contains implementation of a standard payment contract implementation.
+mod constants;
+mod entry_points;
+
+pub use constants::*;
+pub use entry_points::standard_payment_entry_points;
